Move PageObjectInputProvider to its own file to avoid circular imports
