+++ scrapy_poet/utils.py
@@ -13,9 +13,8 @@ from twisted.internet.defer import inlineCallbacks, returnValue
 from web_poet.pages import ItemPage, is_injectable
 from scrapy_poet.providers import (
     providers,
-    PageObjectInputProvider,
 )
-
+from scrapy_poet.page_input_provider import PageObjectInputProvider
 
 _CALLBACK_FOR_MARKER = '__scrapy_poet_callback'
