+++ tests/test_scrapy_dependencies.py
@@ -8,8 +8,8 @@ from web_poet.pages import ItemWebPage
 from scrapy_poet.utils import _SCRAPY_PROVIDED_CLASSES
 from scrapy_poet.providers import (
     provides,
-    PageObjectInputProvider,
 )
+from scrapy_poet.page_input_provider import PageObjectInputProvider
 
 from tests.utils import crawl_items, crawl_single_item, HtmlResource
