+++ tests/test_utils.py
@@ -5,10 +5,10 @@ import scrapy
 from scrapy.http import TextResponse
 
 from scrapy_poet.providers import (
-    PageObjectInputProvider,
     ResponseDataProvider,
     provides,
 )
+from scrapy_poet.page_input_provider import PageObjectInputProvider
 from web_poet.pages import ItemPage, WebPage
 
 from scrapy_poet.utils import (
